requirements.txt
torchvision==0.20.1
tqdm==4.65.0
ffmpeg-python==0.4.1 